+++ src/constants.mjs
@@ -1,0 +1,155 @@
+const CONSTANTS = {
+  MODULE_NAME: "ActiveAuras",
+  SW5E_HUMANOID_RACES: [
+    "abyssin",
+    "aingtii",
+    "aleena",
+    "anzellan",
+    "aqualish",
+    "arcona",
+    "ardennian",
+    "arkanian",
+    "balosar",
+    "barabel",
+    "baragwin",
+    "besalisk",
+    "bith",
+    "bothan",
+    "cathar",
+    "cerean",
+    "chadrafan",
+    "chagrian",
+    "chevin",
+    "chironian",
+    "chiss",
+    "clawdite",
+    "codruji",
+    "colicoid",
+    "dashade",
+    "defel",
+    "devoronian",
+    "draethos",
+    "dug",
+    "duros",
+    "echani",
+    "eshkha",
+    "ewok",
+    "falleen",
+    "felucian",
+    "fleshraider",
+    "gamorrean",
+    "gand",
+    "geonosian",
+    "givin",
+    "gotal",
+    "gran",
+    "gungan",
+    "halfhuman",
+    "harch",
+    "herglic",
+    "ho’din",
+    "human",
+    "hutt",
+    "iktotchi",
+    "ithorian",
+    "jawa",
+    "kage",
+    "kaleesh",
+    "kaminoan",
+    "karkarodon",
+    "keldor",
+    "killik",
+    "klatooinian",
+    "kubaz",
+    "kushiban",
+    "kyuzo",
+    "lannik",
+    "lasat",
+    "lurmen",
+    "miraluka",
+    "mirialan",
+    "moncalamari",
+    "mustafarian",
+    "muun",
+    "nautolan",
+    "neimoidian",
+    "noghri",
+    "ortolan",
+    "patrolian",
+    "pau’an",
+    "pa’lowick",
+    "pyke",
+    "quarren",
+    "rakata",
+    "rattataki",
+    "rishii",
+    "rodian",
+    "ryn",
+    "selkath",
+    "shistavanen",
+    "sithpureblood",
+    "squib",
+    "ssiruu",
+    "sullustan",
+    "talz",
+    "tarasin",
+    "thisspiasian",
+    "togorian",
+    "togruta",
+    "toydarian",
+    "trandoshan",
+    "tusken",
+    "twi'lek",
+    "ugnaught",
+    "umbaran",
+    "verpine",
+    "voss",
+    "vurk",
+    "weequay",
+    "wookie",
+    "yevetha",
+    "zabrak",
+    "zeltron",
+    "zygerrian",
+  ],
+  HUMANOID_RACES: [
+    "human",
+    "orc",
+    "elf",
+    "tiefling",
+    "gnome",
+    "aaracokra",
+    "dragonborn",
+    "dwarf",
+    "halfling",
+    "leonin",
+    "satyr",
+    "genasi",
+    "goliath",
+    "aasimar",
+    "bugbear",
+    "firbolg",
+    "goblin",
+    "lizardfolk",
+    "tabxi",
+    "triton",
+    "yuan-ti",
+    "tortle",
+    "changling",
+    "kalashtar",
+    "shifter",
+    "warforged",
+    "gith",
+    "centaur",
+    "loxodon",
+    "minotaur",
+    "simic hybrid",
+    "vedalken",
+    "verdan",
+    "locathah",
+    "grung",
+  ],
+};
+
+
+export default CONSTANTS;
